+++ maverics/src/main/kotlin/com/example/model/Inventory.kt
@@ -1,6 +1,6 @@
 package com.example.model
 
-class Inventory (freeESOP: Int = 0, lockESOP: Int = 0) {
+class Inventory (freeESOP: Long = 0, lockESOP: Long = 0) {
     var freeESOP = freeESOP;
     var lockESOP = lockESOP;
 }
