+++ maverics/src/main/kotlin/com/example/model/Order.kt
@@ -2,8 +2,8 @@ package com.example.model
 
 class Order {
     var orderId: Int=0
-    var price :Int =0
-    var quantity :Int=0
+    var price : Long=0
+    var quantity :Long=0
     var status :String =""
     var type : String =""
     var userName :String=""
