+++ maverics/src/main/kotlin/com/example/model/OrderHistory.kt
@@ -2,8 +2,8 @@ package com.example.model
 
 data class OrderHistory (
   val orderId: Int,
-  val price :Int ,
-  val quantity :Int,
+  val price :Long ,
+  val quantity :Long,
   val type : String ,
   val filled :MutableList<Transaction>
 )
